requirements.txt
distlib==0.3.8
filelock==3.14.0
pbr==6.0.0
platformdirs==4.2.1
psycopg2==2.8.6
stevedore==5.2.0
django-bootstrap4==24.3
djangorestframework==3.15.2
factory-boy==3.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
Faker==25.9.2
celery==5.4.0
redis==5.0.8
Pillow==10.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
whitenoise==6.7.0
psycopg2==2.9.9
python-dateutil==2.9.0
pytz==2024.1
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
